I LOVE REALISTIC HOI4 ELECTIONS I LOVE REALISTIC HOI4 ELECTIONS I LOVE REALISTIC HOI4 ELECTIONS
I LOVE SPENDING HOURS OF MY LIFE RESEARCHING HOW FOREIGN ELECTIONS WORK AND CODING THEM INTO HOI4 FOR NO OTHER REASON THAN TO SEE IF I CAN BECAUSE NOONE WOULD GIVE A SHIT IF I JUST DID POLISH ELECTIONS BY EVENT OR WHATEVER

I LOVE IT I LOVE IT I LOVE IT I LOVE IT I LOVE IT I LOVE IT I LOVE IT I LOVE IT I LOVE IT I LOVE IT

CRAZY? I WAS CRAZY ONCE. THE LOCKED ME IN A WHITE ROOM. A RUBBER ROOM. WITH A PC AND HOI4. CRAZY? I WAS CRAZY ONCE. THE LOCKED ME IN A WHITE ROOM. A RUBBER ROOM. WITH A PC AND HOI4. CRAZY? I WAS CRAZY ONCE. THE LOCKED ME IN A WHITE ROOM. A RUBBER ROOM. WITH A PC AND HOI4. CRAZY? I WAS CRAZY ONCE. THE LOCKED ME IN A WHITE ROOM. A RUBBER ROOM. WITH A PC AND HOI4. CRAZY? I WAS CRAZY ONCE. THE LOCKED ME IN A WHITE ROOM. A RUBBER ROOM. WITH A PC AND HOI4. CRAZY? I WAS CRAZY ONCE. THE LOCKED ME IN A WHITE ROOM. A RUBBER ROOM. WITH A PC AND HOI4. CRAZY? I WAS CRAZY ONCE. THE LOCKED ME IN A WHITE ROOM. A RUBBER ROOM. WITH A PC AND HOI4. CRAZY? I WAS CRAZY ONCE. THE LOCKED ME IN A WHITE ROOM. A RUBBER ROOM. WITH A PC AND HOI4. CRAZY? I WAS CRAZY ONCE. THE LOCKED ME IN A WHITE ROOM. A RUBBER ROOM. WITH A PC AND HOI4.